[Change] separate public controllers from routes
[Fix] controllers filename
